feat: Add Phase 4 Operational Enhancements documentation and tasks
- Introduced `research.md` detailing key decisions for Phase 4 enhancements including channel isolation, canary rollout strategies, and integrity enforcement.
- Created `tasks.md` outlining actionable tasks grouped by user stories for implementing Phase 4 features, including setup, foundational work, and user stories focused on channel separation, safe promotions, canary exposure, integrity policies, and administrative surfaces.
- Established a comprehensive execution plan with dependencies and parallel opportunities for efficient task management.

chore: Implement architecture and code quality review plan

- Added `plan.md` for a thorough review of the Nuplane solution, identifying 21 actionable items across architecture improvements, code quality, infrastructure enhancements, and testing recommendations.
- Organized recommendations into tiers by priority and execution phases to facilitate systematic improvements.

refactor: Consolidate version handling logic

- Created `NuGetVersionRangeParser.cs` to centralize the logic for extracting concrete version strings from NuGet version range expressions.
- Introduced `VersionKey.cs` to provide a structured representation of version strings, facilitating deterministic version ordering during reconciliation.
- Removed duplicated `VersionKey` and `SelectVersion()` methods from various files, promoting code reuse and maintainability.
