+++ src/Nuplane.Runtime/Reconciliation/DesiredActualDiffEngine.cs
@@ -1,4 +1,5 @@
 using Nuplane.Abstractions;
+using Nuplane.Runtime.Versioning;
 
 namespace Nuplane.Runtime.Reconciliation;
 
@@ -55,70 +56,4 @@ public sealed class DesiredActualDiffEngine
             .ToArray();
     }
 
-    private readonly record struct VersionKey(int Major, int Minor, int Patch, string Suffix) : IComparable<VersionKey>
-    {
-        public static VersionKey Create(string version)
-        {
-            if (string.IsNullOrWhiteSpace(version))
-            {
-                return new(0, 0, 0, string.Empty);
-            }
-
-            var normalized = version.Trim();
-            if (normalized.StartsWith("[", StringComparison.Ordinal) && normalized.EndsWith("]", StringComparison.Ordinal))
-            {
-                normalized = normalized[1..^1];
-            }
-
-            var plusIndex = normalized.IndexOf('+');
-            if (plusIndex >= 0)
-            {
-                normalized = normalized[..plusIndex];
-            }
-
-            var dashIndex = normalized.IndexOf('-');
-            var core = dashIndex >= 0 ? normalized[..dashIndex] : normalized;
-            var suffix = dashIndex >= 0 ? normalized[(dashIndex + 1)..] : string.Empty;
-
-            var coreParts = core.Split('.', StringSplitOptions.RemoveEmptyEntries);
-            _ = int.TryParse(coreParts.ElementAtOrDefault(0), out var major);
-            _ = int.TryParse(coreParts.ElementAtOrDefault(1), out var minor);
-            _ = int.TryParse(coreParts.ElementAtOrDefault(2), out var patch);
-
-            return new(major, minor, patch, suffix);
-        }
-
-        public int CompareTo(VersionKey other)
-        {
-            var majorCompare = Major.CompareTo(other.Major);
-            if (majorCompare != 0)
-            {
-                return majorCompare;
-            }
-
-            var minorCompare = Minor.CompareTo(other.Minor);
-            if (minorCompare != 0)
-            {
-                return minorCompare;
-            }
-
-            var patchCompare = Patch.CompareTo(other.Patch);
-            if (patchCompare != 0)
-            {
-                return patchCompare;
-            }
-
-            if (string.IsNullOrEmpty(Suffix) && !string.IsNullOrEmpty(other.Suffix))
-            {
-                return 1;
-            }
-
-            if (!string.IsNullOrEmpty(Suffix) && string.IsNullOrEmpty(other.Suffix))
-            {
-                return -1;
-            }
-
-            return string.Compare(Suffix, other.Suffix, StringComparison.OrdinalIgnoreCase);
-        }
-    }
 }
