+++ src/Nuplane.Runtime/Versioning/NuGetVersionRangeParser.cs
@@ -1,0 +1,37 @@
+namespace Nuplane.Runtime.Versioning;
+
+/// <summary>
+/// Parses NuGet version range expressions and extracts concrete version strings.
+/// </summary>
+internal static class NuGetVersionRangeParser
+{
+    /// <summary>
+    /// Extracts a concrete version string from a NuGet version range expression.
+    /// For exact versions, returns the version as-is.
+    /// For range expressions (e.g., "[1.0.0, 2.0.0)"), returns the lower bound.
+    /// </summary>
+    public static string SelectVersion(string versionRange)
+    {
+        if (string.IsNullOrWhiteSpace(versionRange))
+        {
+            return "0.0.0";
+        }
+
+        var normalized = versionRange.Trim();
+        if (normalized.StartsWith("[", StringComparison.Ordinal) || normalized.StartsWith("(", StringComparison.Ordinal))
+        {
+            var parts = normalized
+                .TrimStart('[', '(')
+                .TrimEnd(']', ')')
+                .Split(',', StringSplitOptions.TrimEntries | StringSplitOptions.RemoveEmptyEntries);
+
+            if (parts.Length > 0)
+            {
+                return parts[0];
+            }
+        }
+
+        return normalized;
+    }
+}
+
